% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schedule.R
\name{add-schedule}
\alias{add-schedule}
\alias{addSchedule}
\alias{addScheduleDF}
\title{Add schedules}
\usage{
addSchedule(cal, start, end, title, body = NULL, id = NULL,
  calendarId = NULL, category = NULL, ...)

addScheduleDF(cal, data)
}
\arguments{
\item{cal}{A \code{calendar} object.}

\item{start}{The start time.}

\item{end}{The end time.}

\item{title}{The schedule title.}

\item{body}{The schedule body text which is text/plain.}

\item{id}{An id for the schedule.}

\item{calendarId}{An id for the calendar.}

\item{category}{The schedule type ('milestone', 'task', allday', 'time').}

\item{...}{Additionnal arguments passed to the JavaScript method, 
see \url{https://nhn.github.io/tui.calendar/latest/Schedule}.}

\item{data}{a \code{data.frame} with schedule data, columns must have valid shedule name option.}
}
\description{
Add schedules
}
