% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schedule.R
\name{addSchedule}
\alias{addSchedule}
\title{Add schedules}
\usage{
addSchedule(cal, start, end, title, body = NULL, id = NULL,
  calendarId = NULL, category = NULL, ...)
}
\arguments{
\item{cal}{A \code{calendar} object.}

\item{start}{The start time.}

\item{end}{The end time.}

\item{title}{The schedule title.}

\item{body}{The schedule body text which is text/plain.}

\item{id}{An id for the schedule.}

\item{calendarId}{An id for the calendar.}

\item{category}{The schedule type ('milestone', 'task', allday', 'time').}
}
\description{
Add schedules
}
