% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proxy.R
\name{cal_proxy_clear}
\alias{cal_proxy_clear}
\title{Clear calendar with Proxy}
\usage{
cal_proxy_clear(proxy, immediately = TRUE)
}
\arguments{
\item{proxy}{A \code{\link{calendarProxy}} \code{htmlwidget} object.}

\item{immediately}{Render it immediately. Or wait, if you want to add schedule after that for example.}
}
\description{
This function allow to delete all schedules and clear view.
}
