% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proxy.R
\name{cal_proxy_options}
\alias{cal_proxy_options}
\title{Set calendar's options with Proxy}
\usage{
cal_proxy_options(proxy, ...)
}
\arguments{
\item{proxy}{A \code{\link{calendarProxy}} \code{htmlwidget} object.}

\item{...}{Options for the calendar, you can use arguments from \code{\link{calendar}},
\code{\link{setMonthOptions}} (under the form \code{month = list(...)}), 
\code{\link{setWeekOptions}} (under the form \code{week = list(...)})}
}
\description{
This function allow to set options for a calendar.
}
