% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proxy.R
\name{cal_proxy_schedule}
\alias{cal_proxy_schedule}
\title{Add schedule with Proxy}
\usage{
cal_proxy_schedule(proxy, start = NULL, end = NULL, title = NULL,
  body = NULL, id = NULL, calendarId = NULL, category = NULL, ...,
  .list = NULL)
}
\arguments{
\item{proxy}{A \code{\link{calendarProxy}} \code{htmlwidget} object.}

\item{start}{The start time.}

\item{end}{The end time.}

\item{title}{The schedule title.}

\item{body}{The schedule body text which is text/plain.}

\item{id}{An id for the schedule.}

\item{calendarId}{An id for the calendar.}

\item{category}{The schedule type ('milestone', 'task', allday', 'time').}

\item{...}{Additionnal arguments passed to the JavaScript method.}

\item{.list}{A \code{list} with same information as above, useful with \code{input$<outputId>_add_schedule}.}
}
\description{
This function allow to add schedule(s) into a calendar within the server in a Shiny application.
}
