% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proxy.R
\name{cal_proxy_view}
\alias{cal_proxy_view}
\title{Change calendar view with Proxy}
\usage{
cal_proxy_view(proxy, view)
}
\arguments{
\item{proxy}{A \code{\link{calendarProxy}} \code{htmlwidget} object.}

\item{view}{The new view for the calendar: "day", "week" or "month".}
}
\description{
This function allow to change the calendar view from the server in a Shiny application.
}
