% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme.R
\name{set_theme}
\alias{set_theme}
\title{Set theme for a calendar}
\usage{
set_theme(cal, ..., theme_list = NULL)
}
\arguments{
\item{cal}{A \code{calendar} object.}

\item{...}{Named arguments to customize appearance with CSS. See online documentation for full list of options.}

\item{theme_list}{Alternative to \code{...} for using a list.}
}
\description{
Full configuration for theme.
"common" prefix is for entire calendar. "common" properties can be overriden by "week", "month".
"week" prefix is for weekly and daily view.
"month" prefix is for monthly view.
}
\note{
Online JavaScript documentation: \url{https://nhn.github.io/tui.calendar/latest/themeConfig}
}
\examples{
calendar(defaultView = "month") \%>\%
  set_theme(
    common.border = '2px solid #e5e5e5',
    common.backgroundColor = 'white',
    common.holiday.color = '#ff4040',
    common.saturday.color = '#0000FF',
    common.dayname.color = '#333',
    common.today.color = '#333'
  )
}
