% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set-calendars.R
\name{setCalendarsProps}
\alias{setCalendarsProps}
\title{Calendars properties}
\usage{
setCalendarsProps(cal, id, name, color, bgColor, borderColor = bgColor)
}
\arguments{
\item{cal}{A \code{calendar} object.}

\item{id}{Id for the calendar, refer to \code{calendarId} in \code{\link{addSchedule}}}

\item{name}{The calendar name.}

\item{color}{The text color when schedule is displayed.}

\item{bgColor}{The background color schedule is displayed.}

\item{borderColor}{The color of left border or bullet point when schedule is displayed.}
}
\description{
Define calendadrs properties for schedules
}
