% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set-options.R
\name{setWeekOptions}
\alias{setWeekOptions}
\title{Week Options}
\usage{
setWeekOptions(cal, startDayOfWeek = NULL, daynames = NULL,
  narrowWeekend = NULL, workweek = NULL,
  showTimezoneCollapseButton = NULL, timezonesCollapsed = NULL,
  hourStart = NULL, hourEnd = NULL)
}
\arguments{
\item{cal}{A \code{calendar} object.}

\item{startDayOfWeek}{Numeric. The start day of week.}

\item{daynames}{Vector. The day names in weekly and daily. Default values are 'Sun', 'Mon', 'Tue', 'Wed', 'Thu', 'Fri', 'Sat'.}

\item{narrowWeekend}{Logical. Make weekend column narrow(1/2 width).}

\item{workweek}{Logical. Show only 5 days except for weekend.}

\item{showTimezoneCollapseButton}{Logical. Show a collapse button to close multiple timezones}

\item{timezonesCollapsed}{Logical. An initial multiple timezones collapsed state.}

\item{hourStart}{Numeric. Can limit of render hour start.}

\item{hourEnd}{Numeric. Can limit of render hour end.}
}
\description{
Options for daily, weekly view.
}
\note{
Online JavaScript documentation: \url{https://nhn.github.io/tui.calendar/latest/WeekOptions}
}
