% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\name{tuicalendr-exports}
\alias{tuicalendr-exports}
\alias{\%>\%}
\alias{JS}
\title{tuicalendr exported operators and S3 methods}
\description{
The following functions are imported and then re-exported
from the tuicalendr package to avoid listing the magrittr
as Depends of tuicalendr
}
