% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{tuicalendr-package}
\alias{tuicalendr-package}
\title{An \code{htmlwidget} interface to the
tui-calendar javascript library}
\description{
This package allow you to use tui-calendar.js (\url{https://github.com/nhn/tui.calendar}),
to create interactive calendars.
}
\author{
Victor Perrier (@dreamRs_fr)
}
