% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proxy.R
\name{cal_proxy_view}
\alias{cal_proxy_view}
\title{Change calendar view with Proxy}
\usage{
cal_proxy_view(proxy, view)
}
\arguments{
\item{proxy}{A \code{\link{calendarProxy}} \code{htmlwidget} object.}

\item{view}{The new view for the calendar: "day", "week" or "month".}
}
\description{
This function allow to change the calendar view from the server in a Shiny application.
}
\examples{
if (interactive()) {
  library(shiny)
  
  ui <- fluidPage(
    tags$h2("Navigate in calendar with actionButtons"),
    radioButtons(
      inputId = "view", 
      label = "Change view:", 
      choices = c("day", "week", "month"), 
      inline = TRUE
    ),
    calendarOutput(outputId = "my_calendar")
  )
  
  server <- function(input, output, session) {
    
    output$my_calendar <- renderCalendar({
      calendar(defaultView = "day", scheduleView = "allday") \%>\% 
        addSchedule(
          title = "Today planning", 
          start = Sys.Date(),
          end = Sys.Date(),
          category = "allday"
        )
    })
    
    observeEvent(
      input$view, 
      cal_proxy_view("my_calendar", input$view), 
      ignoreInit = TRUE
    )
    
  }
  
  shinyApp(ui, server)
}
}
