% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proxy.R
\name{proxy-navigate}
\alias{proxy-navigate}
\alias{cal_proxy_next}
\alias{cal_proxy_prev}
\alias{cal_proxy_today}
\title{Navigate into a calendar with Proxy}
\usage{
cal_proxy_next(proxy)

cal_proxy_prev(proxy)

cal_proxy_today(proxy)
}
\arguments{
\item{proxy}{A \code{\link{calendarProxy}} \code{htmlwidget} object.}
}
\description{
Those functions allow to navigate in the calendar from the server in a Shiny application.
}
