% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set-calendars.R
\name{set-calendars}
\alias{set-calendars}
\alias{set_calendars_props}
\alias{set_calendars_props_df}
\title{Calendars properties}
\usage{
set_calendars_props(cal, id, name, color, bgColor, borderColor = bgColor)

set_calendars_props_df(cal, df)
}
\arguments{
\item{cal}{A \code{calendar} object.}

\item{id}{Id for the calendar, refer to \code{calendarId} in \code{\link{add_schedule}}}

\item{name}{The calendar name.}

\item{color}{The text color when schedule is displayed.}

\item{bgColor}{The background color schedule is displayed.}

\item{borderColor}{The color of left border or bullet point when schedule is displayed.}

\item{df}{a \code{data.frame} with calendars data, columns must have valid calendars properties option.}
}
\description{
Define calendars properties for schedules
}
